feat: move VueLanguageServer class in a dedicate file
